Refactor aggregate export.
