+++ report-processor/src/main/java/org/opentestsystem/rdw/reporting/processor/web/converter/AbstractAggregateReportCsvHandler.java
@@ -12,10 +12,13 @@ import org.apache.commons.csv.CSVPrinter;
 import org.opentestsystem.rdw.reporting.common.model.AbstractAggregateReportQuery;
 import org.opentestsystem.rdw.reporting.common.model.AggregateRow;
 import org.opentestsystem.rdw.reporting.common.model.Dimension;
+import org.opentestsystem.rdw.reporting.common.model.DimensionType;
 import org.opentestsystem.rdw.reporting.common.model.Measures;
 import org.opentestsystem.rdw.reporting.common.model.StudentFilters;
 import org.opentestsystem.rdw.reporting.common.util.Headers;
+import org.opentestsystem.rdw.reporting.processor.model.AbstractExamReportRequest;
 import org.opentestsystem.rdw.reporting.processor.model.AggregateReportRequest;
+import org.opentestsystem.rdw.reporting.processor.model.ReportType;
 import org.opentestsystem.rdw.reporting.processor.web.ReportContentResource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -23,6 +26,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.MessageSource;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpOutputMessage;
+import org.springframework.http.MediaType;
 import org.springframework.http.converter.HttpMessageNotWritableException;
 
 import javax.validation.constraints.NotNull;
@@ -44,11 +48,13 @@ import java.util.concurrent.atomic.AtomicInteger;
 import static com.google.common.collect.Lists.newArrayList;
 import static com.google.common.collect.Maps.newHashMap;
 import static org.apache.commons.lang.StringUtils.EMPTY;
+import static org.apache.commons.lang.StringUtils.isNotEmpty;
 import static org.apache.commons.lang.StringUtils.trimToEmpty;
 import static org.opentestsystem.rdw.common.model.AssessmentType.IAB;
 import static org.opentestsystem.rdw.reporting.common.util.MediaTypes.TEXT_CSV_UTF8;
 import static org.opentestsystem.rdw.reporting.common.util.MediaTypes.getExtension;
 import static org.springframework.http.HttpHeaders.CONTENT_DISPOSITION;
+import static org.springframework.http.MediaType.APPLICATION_JSON;
 
 /**
  * This base handler provides common functionality for converting reports consisting of a list of
@@ -58,8 +64,8 @@ import static org.springframework.http.HttpHeaders.CONTENT_DISPOSITION;
  * last column of the response under the header of "Attributes."
  */
 @SuppressWarnings("unchecked")
-abstract class AggregateReportCsvHandler<Q extends AbstractAggregateReportQuery, R extends AggregateRow> extends DefaultReportHandler {
-    private static final Logger logger = LoggerFactory.getLogger(AggregateReportCsvHandler.class);
+abstract class AbstractAggregateReportCsvHandler<Q extends AbstractAggregateReportQuery, R extends AggregateRow> extends DefaultReportHandler {
+    private static final Logger logger = LoggerFactory.getLogger(AbstractAggregateReportCsvHandler.class);
     private static final String NotApplicable = "";
     private static final Object[] EmptyObjectArray = new Object[]{};
     private static final Locale ReportLocale = Locale.US;
@@ -71,10 +77,17 @@ abstract class AggregateReportCsvHandler<Q extends AbstractAggregateReportQuery,
     private final ObjectMapper objectMapper;
 
     @Autowired
-    public AggregateReportCsvHandler(final MessageSource messageSource,
-                                     final ObjectMapper objectMapper) {
+    public AbstractAggregateReportCsvHandler(final MessageSource messageSource,
+                                             final ObjectMapper objectMapper) {
         this.messageSource = messageSource;
         this.objectMapper = objectMapper;
+    }
+
+    @Override
+    public boolean accept(final ReportContentResource resource, final MediaType requestedMediaType) {
+        final AbstractExamReportRequest request = resource.getReport().getReportRequest();
+        return !APPLICATION_JSON.isCompatibleWith(requestedMediaType) &&
+                request.getReportType() == ReportType.AggregateReportRequest;
     }
 
     @Override
@@ -183,27 +196,52 @@ abstract class AggregateReportCsvHandler<Q extends AbstractAggregateReportQuery,
     /**
      * Returns query data that need to be appended as attributes to the first row
      */
-    protected List<String> getFirstRowAttributes(final Q query) {
+    private List<String> getFirstRowAttributes(final Q query) {
         final List<String> attributes = newArrayList();
 
         final LocalDateTime localDate = LocalDateTime.now();
         appendAttribute(attributes, "ReportDate", Collections.singleton(localDate.format(DateTimeFormatter.ISO_DATE_TIME)));
         appendAttribute(attributes, "AdministrationCondition", query.getAdministrativeConditionCodes());
         appendAttribute(attributes, "Completeness", query.getCompletenessCodes());
+        if (!query.isFilteredSubgroupQuery()) attributes.addAll(getFiltersAsAttributes(query.getStudentFilters()));
         return attributes;
+    }
+
+    private String getDimensionValue(final CsvContext context, final Dimension dimension) {
+        return !context.getQuery().isFilteredSubgroupQuery() ? trimToEmpty(dimension.getCode()) : getFilteredSubgroupDimensionValue(context, dimension);
+    }
+
+    /**
+     * Returns a custom subgroup {@link StudentFilters} as a string of attributes
+     * A subgroup is identified by the {@link Dimension#code}
+     */
+    private String getFilteredSubgroupDimensionValue(final CsvContext context, final Dimension dimension) {
+
+        final String subgroupKey = dimension.getCode();
+
+        //check the cached values first
+        if (isNotEmpty(context.get(subgroupKey, String.class))) return context.get(subgroupKey, String.class);
+
+        final String subgroupAttributesAsString = getSubgroupAsString(context, subgroupKey);
+        context.put(subgroupKey, subgroupAttributesAsString);
+        return subgroupAttributesAsString;
+    }
+
+    private String getSubgroupAsString(final CsvContext context, final String subgroupKey) {
+
+        final StudentFilters subGroupStudentFilters = context.getQuery().getSubgroups().get(subgroupKey);
+        if (subGroupStudentFilters != null) return AttributeJoiner.join(getFiltersAsAttributes(subGroupStudentFilters));
+
+        // check if this is a default Overall dimension that was added without a user's explicit request
+        if (DimensionType.Overall.equals(DimensionType.caseInsensitiveValue(subgroupKey))) return DimensionType.Overall.code();
+
+        throw new IllegalStateException("Unknown subgroup referenced by FilteredSubgroup report result: " + subgroupKey);
     }
 
     /**
      * @return a class of a row instance
      */
     protected abstract Class<R> getRowInstanceClass();
-
-    /**
-     * @return a dimension's value. This allows custom implementations to override the default behaviour
-     */
-    protected String getDimensionValue(final CsvContext context, final Dimension dimension) {
-        return trimToEmpty(dimension.getCode());
-    }
 
     /**
      * @return a list of all {@link Measures} headers. This allows custom implementations to override the default behaviour
