+++ report-processor/src/main/java/org/opentestsystem/rdw/reporting/processor/web/converter/CustomAggregateReportCsvHandler.java
@@ -15,7 +15,7 @@ import java.util.List;
 import static org.opentestsystem.rdw.reporting.common.model.AggregateReportType.CustomAggregate;
 
 @Component
-public class CustomAggregateReportCsvHandler extends BasicAggregateReportCsvHandler<AggregateRow, CustomAggregateReportQuery> {
+public class CustomAggregateReportCsvHandler extends AbstractAggregateReportCsvHandler<CustomAggregateReportQuery, AggregateRow> {
 
     public CustomAggregateReportCsvHandler(final MessageSource messageSource, final ObjectMapper objectMapper) {
         super(messageSource, objectMapper);
@@ -27,7 +27,7 @@ public class CustomAggregateReportCsvHandler extends BasicAggregateReportCsvHand
         if (!accept) return false;
 
         final AggregateQuery query = ((AggregateReportRequest) resource.getReport().getReportRequest()).getQuery();
-        return query.getReportType() == CustomAggregate && !query.isFilteredSubgroupQuery();
+        return query.getReportType() == CustomAggregate;
     }
 
     @Override
