--- report-processor/src/main/java/org/opentestsystem/rdw/reporting/processor/web/converter/CustomFilteredSubgroupReportCsvHandler.java
@@ -1,47 +1,0 @@
-package org.opentestsystem.rdw.reporting.processor.web.converter;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.opentestsystem.rdw.reporting.common.model.AggregateQuery;
-import org.opentestsystem.rdw.reporting.common.model.AggregateRow;
-import org.opentestsystem.rdw.reporting.common.model.CustomAggregateReportQuery;
-import org.opentestsystem.rdw.reporting.processor.model.AggregateReportRequest;
-import org.opentestsystem.rdw.reporting.processor.web.ReportContentResource;
-import org.springframework.context.MessageSource;
-import org.springframework.http.MediaType;
-import org.springframework.stereotype.Component;
-
-import java.util.List;
-
-import static org.opentestsystem.rdw.reporting.common.model.AggregateReportType.CustomAggregate;
-
-@Component
-public class CustomFilteredSubgroupReportCsvHandler extends FilteredSubgroupReportCsvHandler<AggregateRow, CustomAggregateReportQuery> {
-
-    public CustomFilteredSubgroupReportCsvHandler(final MessageSource messageSource, final ObjectMapper objectMapper) {
-        super(messageSource, objectMapper);
-    }
-
-    @Override
-    public boolean accept(final ReportContentResource resource, final MediaType requestedMediaType) {
-        final boolean accept = super.accept(resource, requestedMediaType);
-        if (!accept) return false;
-
-        final AggregateQuery query = ((AggregateReportRequest) resource.getReport().getReportRequest()).getQuery();
-        return query.getReportType() == CustomAggregate && query.isFilteredSubgroupQuery();
-    }
-
-    @Override
-    protected List<String> getMeasureHeaders(final CustomAggregateReportQuery query) {
-        return getMeasureHeaders(query, "");
-    }
-
-    @Override
-    protected List<String> getMeasuresValues(final CsvContext context, final AggregateRow result) {
-        return getMeasuresValues(context.getQuery().getAssessmentTypeCode(), result.getMeasures());
-    }
-
-    @Override
-    protected Class<AggregateRow> getRowInstanceClass() {
-        return AggregateRow.class;
-    }
-}
