+++ report-processor/src/main/java/org/opentestsystem/rdw/reporting/processor/web/converter/LongitudinalAggregateReportCsvHandler.java
@@ -15,7 +15,7 @@ import java.util.List;
 import static org.opentestsystem.rdw.reporting.common.model.AggregateReportType.Longitudinal;
 
 @Component
-public class LongitudinalAggregateReportCsvHandler extends BasicAggregateReportCsvHandler<LongitudinalRow, LongitudinalReportQuery> {
+public class LongitudinalAggregateReportCsvHandler extends AbstractAggregateReportCsvHandler<LongitudinalReportQuery, LongitudinalRow> {
     public LongitudinalAggregateReportCsvHandler(final MessageSource messageSource, final ObjectMapper objectMapper) {
         super(messageSource, objectMapper);
     }
@@ -26,7 +26,7 @@ public class LongitudinalAggregateReportCsvHandler extends BasicAggregateReportC
         if (!accept) return false;
 
         final AggregateQuery query = ((AggregateReportRequest) resource.getReport().getReportRequest()).getQuery();
-        return query.getReportType() == Longitudinal && !query.isFilteredSubgroupQuery();
+        return query.getReportType() == Longitudinal;
     }
 
     @Override
