+++ report-processor/src/test/java/org/opentestsystem/rdw/reporting/processor/web/converter/AbstractAggregateReportCsvHandlerTest.java
@@ -11,7 +11,7 @@ import org.opentestsystem.rdw.reporting.common.model.State;
 
 import java.util.concurrent.atomic.AtomicInteger;
 
-public abstract class AggregateReportCsvHandlerTest {
+public abstract class AbstractAggregateReportCsvHandlerTest {
     private static final AtomicInteger idGenerator = new AtomicInteger(1);
 
     protected State state() {
