+++ report-processor/src/test/java/org/opentestsystem/rdw/reporting/processor/web/converter/AbstractAggregateReportCsvHandlerWithBasicQueryTest.java
@@ -39,7 +39,7 @@ import static org.opentestsystem.rdw.reporting.common.util.MediaTypes.TEXT_CSV_U
 import static org.springframework.http.MediaType.APPLICATION_JSON;
 
 
-public abstract class BasicAggregateReportCsvHandlerTest<R extends AggregateRow> extends AggregateReportCsvHandlerTest {
+public abstract class AbstractAggregateReportCsvHandlerWithBasicQueryTest<R extends AggregateRow> extends AbstractAggregateReportCsvHandlerTest {
 
     @Mock
     private MessageSource messageSource;
@@ -50,11 +50,11 @@ public abstract class BasicAggregateReportCsvHandlerTest<R extends AggregateRow>
     private MockHttpOutputMessage outputMessage;
     private byte[] reportPayload;
     private Report report;
-    private BasicAggregateReportCsvHandler handler;
+    private AbstractAggregateReportCsvHandler handler;
 
     protected abstract R reportRow(final Organization organization);
 
-    protected abstract BasicAggregateReportCsvHandler createHandlerUnderTest(MessageSource messageSource, ObjectMapper objectMapper);
+    protected abstract AbstractAggregateReportCsvHandler createHandlerUnderTest(MessageSource messageSource, ObjectMapper objectMapper);
 
     protected abstract AbstractAggregateReportQuery.Builder getQueryBuilder();
