+++ report-processor/src/test/java/org/opentestsystem/rdw/reporting/processor/web/converter/AbstractAggregateReportCsvHandlerWithFilteredSubgroupQueryTest.java
@@ -40,7 +40,7 @@ import static org.mockito.Mockito.when;
 import static org.opentestsystem.rdw.reporting.common.util.MediaTypes.TEXT_CSV_UTF8;
 import static org.springframework.http.MediaType.APPLICATION_JSON;
 
-abstract class FilteredSubgroupReportCsvHandlerTest<R extends AggregateRow> extends AggregateReportCsvHandlerTest {
+abstract class AbstractAggregateReportCsvHandlerWithFilteredSubgroupQueryTest<R extends AggregateRow> extends AbstractAggregateReportCsvHandlerTest {
 
     @Mock
     private MessageSource messageSource;
@@ -51,11 +51,11 @@ abstract class FilteredSubgroupReportCsvHandlerTest<R extends AggregateRow> exte
     private MockHttpOutputMessage outputMessage;
     private byte[] reportPayload;
     private Report report;
-    private FilteredSubgroupReportCsvHandler handler;
+    private AbstractAggregateReportCsvHandler handler;
 
     protected abstract R reportRow(String subgroupKey, Organization organization);
 
-    protected abstract FilteredSubgroupReportCsvHandler createHandlerUnderTest(MessageSource messageSource, ObjectMapper objectMapper);
+    protected abstract AbstractAggregateReportCsvHandler createHandlerUnderTest(MessageSource messageSource, ObjectMapper objectMapper);
 
     protected abstract AbstractAggregateReportQuery.Builder getQueryBuilder();
