+++ report-processor/src/test/java/org/opentestsystem/rdw/reporting/processor/web/converter/CustomAggregateReportCsvHandlerWithBasicQueryTest.java
@@ -13,7 +13,7 @@ import org.springframework.context.MessageSource;
 import static org.opentestsystem.rdw.reporting.common.model.DimensionType.Overall;
 
 @RunWith(MockitoJUnitRunner.class)
-public class CustomBasicAggregateReportCsvHandlerTest extends BasicAggregateReportCsvHandlerTest<AggregateRow> {
+public class CustomAggregateReportCsvHandlerWithBasicQueryTest extends AbstractAggregateReportCsvHandlerWithBasicQueryTest<AggregateRow> {
 
     @Override
     protected AggregateRow reportRow(final Organization organization) {
@@ -23,7 +23,7 @@ public class CustomBasicAggregateReportCsvHandlerTest extends BasicAggregateRepo
     }
 
     @Override
-    protected BasicAggregateReportCsvHandler createHandlerUnderTest(final MessageSource messageSource, final ObjectMapper objectMapper) {
+    protected CustomAggregateReportCsvHandler createHandlerUnderTest(final MessageSource messageSource, final ObjectMapper objectMapper) {
         return new CustomAggregateReportCsvHandler(messageSource, objectMapper);
     }
