+++ report-processor/src/test/java/org/opentestsystem/rdw/reporting/processor/web/converter/CustomAggregateReportCsvHandlerWithFilteredSubgroupQueryTest.java
@@ -12,7 +12,7 @@ import org.opentestsystem.rdw.reporting.common.test.support.ReportQueries;
 import org.springframework.context.MessageSource;
 
 @RunWith(MockitoJUnitRunner.class)
-public class CustomFilteredSubgroupReportCsvHandlerTest extends FilteredSubgroupReportCsvHandlerTest<AggregateRow> {
+public class CustomAggregateReportCsvHandlerWithFilteredSubgroupQueryTest extends AbstractAggregateReportCsvHandlerWithFilteredSubgroupQueryTest<AggregateRow> {
 
     @Override
     protected AggregateRow reportRow(final String subgroupKey, final Organization organization) {
@@ -22,8 +22,8 @@ public class CustomFilteredSubgroupReportCsvHandlerTest extends FilteredSubgroup
     }
 
     @Override
-    protected FilteredSubgroupReportCsvHandler createHandlerUnderTest(final MessageSource messageSource, final ObjectMapper objectMapper) {
-        return new CustomFilteredSubgroupReportCsvHandler(messageSource, objectMapper);
+    protected CustomAggregateReportCsvHandler createHandlerUnderTest(final MessageSource messageSource, final ObjectMapper objectMapper) {
+        return new CustomAggregateReportCsvHandler(messageSource, objectMapper);
     }
 
     @Override
