+++ report-processor/src/test/java/org/opentestsystem/rdw/reporting/processor/web/converter/LongitudinalAggregateReportCsvHandlerWithBasicQueryTest.java
@@ -11,52 +11,39 @@ import org.opentestsystem.rdw.reporting.common.model.LongitudinalReportQuery;
 import org.opentestsystem.rdw.reporting.common.model.LongitudinalRow;
 import org.opentestsystem.rdw.reporting.common.model.Measures;
 import org.opentestsystem.rdw.reporting.common.model.Organization;
-import org.opentestsystem.rdw.reporting.common.model.StudentFilters;
 import org.springframework.context.MessageSource;
 
-import java.util.Map;
-
-import static com.google.common.collect.Maps.newHashMap;
-import static org.opentestsystem.rdw.common.model.AssessmentType.ICA;
+import static org.opentestsystem.rdw.common.model.AssessmentType.SUMMATIVE;
 import static org.opentestsystem.rdw.common.model.Subject.MATH;
-import static org.opentestsystem.rdw.reporting.common.model.DimensionType.Custom;
-
+import static org.opentestsystem.rdw.reporting.common.model.DimensionType.Gender;
+import static org.opentestsystem.rdw.reporting.common.model.DimensionType.Overall;
 
 @RunWith(MockitoJUnitRunner.class)
-public class LongitudinalFilteredSubgroupReportCsvHandlerTest extends FilteredSubgroupReportCsvHandlerTest<LongitudinalRow> {
+public class LongitudinalAggregateReportCsvHandlerWithBasicQueryTest extends AbstractAggregateReportCsvHandlerWithBasicQueryTest<LongitudinalRow> {
+
     @Override
-    protected LongitudinalRow reportRow(final String subgroupKey, final Organization organization) {
+    protected LongitudinalRow reportRow(final Organization organization) {
         return super.reportRowBuilder(LongitudinalRow.builder(), organization)
-                .dimension(Dimension.builder().type(Custom).code("subgroupA").build())
+                .dimension(Dimension.builder().type(Overall).build())
                 .cohortMeasures(Measures.builder().build())
                 .build();
     }
 
     @Override
-    protected FilteredSubgroupReportCsvHandler createHandlerUnderTest(final MessageSource messageSource, final ObjectMapper objectMapper) {
-        return new LongitudinalFilteredSubgroupReportCsvHandler(messageSource, objectMapper);
+    protected LongitudinalAggregateReportCsvHandler createHandlerUnderTest(final MessageSource messageSource, final ObjectMapper objectMapper) {
+        return new LongitudinalAggregateReportCsvHandler(messageSource, objectMapper);
     }
 
     @Override
     protected AbstractAggregateReportQuery.Builder getQueryBuilder() {
-        final Map<String, StudentFilters> subgroups = newHashMap();
-        subgroups.put("overall", StudentFilters.builder()
-                .build());
-        subgroups.put("subgroupA", StudentFilters.builder()
-                .genderCodes(ImmutableSet.of("Male"))
-                .build());
-        subgroups.put("subgroupB", StudentFilters.builder()
-                .genderCodes(ImmutableSet.of("Female"))
-                .build());
-
         return LongitudinalReportQuery.builder()
-                .assessmentTypeCode(ICA.code())
+                .assessmentTypeCode(SUMMATIVE.code())
                 .subjectCodes(ImmutableSet.of(MATH.code().toLowerCase(), Subject.ELA.code().toUpperCase()))
                 .toSchoolYear(2018)
                 .assessmentGradeCodes(ImmutableSet.of("01", "02", "03", "04", "05"))
                 .includeState(true)
                 .includeAllDistricts(false)
-                .subgroups(subgroups)
+                .dimensionTypes(ImmutableSet.of(Gender))
                 .districtIds(ImmutableSet.of(1L))
                 .schoolIds(ImmutableSet.of(1L));
     }
